requirements.txt
fastapi
uvicorn[standard]
python-multipart

torch
open-clip-torch

faiss-cpu
numpy
Pillow
